requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
asyncio==3.4.3
attrs==21.2.0
certifi==2020.12.5
chardet==4.0.0
cycler==0.10.0
discord==1.0.1
discord.py==1.7.2
idna==2.10
kiwisolver==1.3.1
matplotlib==3.4.2
multidict==5.1.0
nepse==0.1.2
numpy==1.20.3
pandas==1.2.4
Pillow==8.2.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.17.1
pytz==2021.1
requests==2.25.1
six==1.16.0
typing-extensions==3.10.0.0
urllib3==1.26.4
yarl==1.6.3
